refactor: actualizar el componente y el servicio de reportes para utilizar el modelo de reportes actualizado
