+++ client/src/app/_components/reportes/reportes-page.component.css
@@ -18,13 +18,3 @@ table th {
   background-color: #fff8e1;
   font-weight: 600;
 }
-
-@keyframes spin {
-  to {
-    transform: rotate(360deg);
-  }
-}
-
-.animate-spin {
-  animation: spin 1s linear infinite;
-}
