+++ client/src/app/_models/reporte.model.ts
